refactoring h01.py and microns.py / add example notebook for h01 interface
